(PC-18448)[PRO] fix: separate route and subroutes to avoid import error in tests
