+++ pro/src/app/AppRouter/routes_map.tsx
@@ -28,31 +28,16 @@ import CollectiveDataEdition from 'pages/Offerers/Offerer/VenueV1/VenueEdition/C
 import OffererCreation from 'pages/Offerers/OffererCreation'
 import { OffererStats } from 'pages/OffererStats'
 import { OfferIndividualWizard } from 'pages/OfferIndividualWizard'
-import { Confirmation } from 'pages/OfferIndividualWizard/Confirmation'
-import { Offer } from 'pages/OfferIndividualWizard/Offer'
-import { PriceCategories } from 'pages/OfferIndividualWizard/PriceCategories'
-import { Stocks } from 'pages/OfferIndividualWizard/Stocks'
-import { Summary } from 'pages/OfferIndividualWizard/Summary'
 import OffersRoute from 'pages/Offers'
 import OfferType from 'pages/OfferType'
 import Reimbursements from 'pages/Reimbursements'
 import SignIn from 'pages/SignIn/SignIn'
 import Signup from 'pages/Signup/Signup'
-import SignupConfirmation from 'pages/Signup/SignupConfirmation/SignupConfirmation'
-import SignupContainer from 'pages/Signup/SignupContainer/SignupContainer'
-import SignUpValidation from 'pages/Signup/SignUpValidation'
 import { SignupJourneyRoutes } from 'pages/SignupJourneyRoutes'
 import { UserProfile } from 'pages/User'
 import { VenueCreation } from 'pages/VenueCreation'
 import { VenueEdition } from 'pages/VenueEdition'
 import CollectiveOfferSelectionDuplication from 'screens/CollectiveOfferSelectionDuplication'
-import { Activity } from 'screens/SignupJourneyForm/Activity'
-import { OffererAuthentication } from 'screens/SignupJourneyForm/Authentication'
-import { ConfirmedAttachment } from 'screens/SignupJourneyForm/ConfirmedAttachment'
-import { Offerer } from 'screens/SignupJourneyForm/Offerer'
-import { Offerers as SignupJourneyOfferers } from 'screens/SignupJourneyForm/Offerers'
-import { Validation } from 'screens/SignupJourneyForm/Validation'
-import { Welcome } from 'screens/SignupJourneyForm/Welcome'
 import { UNAVAILABLE_ERROR_PAGE } from 'utils/routes'
 
 export interface ILayoutConfig {
@@ -347,178 +332,10 @@ const routes: IRoute[] = [
   },
 ]
 
-export const routesOfferIndividualWizard: IRoute[] = [
-  {
-    element: <Offer />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/informations',
-    title: 'Détails de l’offre',
-  },
-  {
-    element: <Offer />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/creation/informations',
-    title: 'Détails de l’offre',
-  },
-  {
-    element: <Offer />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/brouillon/informations',
-    title: 'Détails de l’offre',
-  },
-  {
-    element: <PriceCategories />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/creation/tarifs',
-    title: 'Vos tarifs',
-  },
-  {
-    element: <PriceCategories />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/brouillon/tarifs',
-    title: 'Vos tarifs',
-  },
-  {
-    element: <PriceCategories />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/tarifs',
-    title: 'Vos tarifs',
-  },
-  {
-    element: <Stocks />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/creation/stocks',
-    title: 'Vos stocks',
-  },
-  {
-    element: <Stocks />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/brouillon/stocks',
-    title: 'Vos stocks',
-  },
-  {
-    element: <Stocks />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/stocks',
-    title: 'Vos stocks',
-  },
-  {
-    element: <Summary />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/creation/recapitulatif',
-    title: 'Récapitulatif',
-  },
-  {
-    element: <Summary />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/brouillon/recapitulatif',
-    title: 'Récapitulatif',
-  },
-  {
-    element: <Summary />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/recapitulatif',
-    title: 'Récapitulatif',
-  },
-  {
-    element: <Confirmation />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/creation/confirmation',
-    title: 'Confirmation',
-  },
-  {
-    element: <Confirmation />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/brouillon/confirmation',
-    title: 'Confirmation',
-  },
-  {
-    element: <Confirmation />,
-    parentPath: '/offre/individuelle/:offerId',
-    path: '/confirmation',
-    title: 'Confirmation',
-  },
-]
-
-export const routesSignup: IRoute[] = [
-  {
-    element: <SignupContainer />,
-    parentPath: '/inscription',
-    path: '/',
-    title: 'S’inscrire',
-  },
-  {
-    element: <SignupConfirmation />,
-    parentPath: '/inscription',
-    path: '/confirmation',
-    title: 'S’inscrire',
-  },
-  {
-    element: <SignUpValidation />,
-    parentPath: '/inscription',
-    path: '/validation/:token',
-    title: 'S’inscrire',
-  },
-]
-
-export const routesSignupJourney: IRoute[] = [
-  {
-    element: <Welcome />,
-    parentPath: '/parcours-inscription',
-    path: '/',
-    title: '',
-  },
-  {
-    element: <Offerer />,
-    parentPath: '/parcours-inscription',
-    path: '/structure',
-    title: '',
-  },
-  {
-    element: <SignupJourneyOfferers />,
-    parentPath: '/parcours-inscription',
-    path: '/structure/rattachement',
-    title: '',
-  },
-  {
-    element: <ConfirmedAttachment />,
-    parentPath: '/parcours-inscription',
-    path: '/structure/rattachement/confirmation',
-    title: '',
-  },
-  {
-    element: <OffererAuthentication />,
-    parentPath: '/parcours-inscription',
-    path: '/authentification',
-    title: '',
-  },
-  {
-    element: <OffererAuthentication />,
-    parentPath: '/parcours-inscription',
-    path: '/authentification',
-    title: '',
-  },
-  {
-    element: <Activity />,
-    parentPath: '/parcours-inscription',
-    path: '/activite',
-    title: '',
-  },
-  {
-    element: <Validation />,
-    parentPath: '/parcours-inscription',
-    path: '/validation',
-    title: '',
-  },
-]
-
-export const routesDefinitions: RouteDefinition[] = [
-  ...routes,
-  ...routesOfferIndividualWizard,
-  ...routesSignup,
-  ...routesSignupJourney,
-].map(({ path, parentPath, title }) => {
-  return { path, parentPath, title }
-})
+export const routesDefinitions: RouteDefinition[] = routes.map(
+  ({ path, parentPath, title }) => {
+    return { path, parentPath, title }
+  }
+)
 
 export default routes
