+++ pro/src/pages/Signup/Signup.tsx
@@ -2,7 +2,8 @@ import React, { useEffect } from 'react'
 import { Routes, Route } from 'react-router-dom'
 
 import AppLayout from 'app/AppLayout'
-import { routesSignup, IRoute } from 'app/AppRouter/routes_map'
+import { IRoute } from 'app/AppRouter/routes_map'
+import { routesSignup } from 'app/AppRouter/subroutes_map'
 import PageTitle from 'components/PageTitle/PageTitle'
 import useActiveFeature from 'hooks/useActiveFeature'
 import { campaignTracker } from 'tracking/mediaCampaignsTracking'
