+++ pro/src/pages/SignupJourneyRoutes/SignupJourneyRoutes.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { Route, Routes, useLocation, useNavigate } from 'react-router-dom'
 
 import AppLayout from 'app/AppLayout'
-import { routesSignupJourney, IRoute } from 'app/AppRouter/routes_map'
+import { IRoute } from 'app/AppRouter/routes_map'
+import { routesSignupJourney } from 'app/AppRouter/subroutes_map'
 import { SignupJourneyFormLayout } from 'components/SignupJourneyFormLayout'
 import { SignupJourneyContextProvider } from 'context/SignupJourneyContext'
 import { Events } from 'core/FirebaseEvents/constants'
